Update README.md
The objective of this assignment is to get you comfortable with numbering systems, bitwise operations, and common binary operations. The programming assignment should be implemented in Java. You are required to work alone on this assignment. The assignment accounts for 7.5% towards your cumulative course grade.

This assignment may be modified to clarify any questions (and the version number incremented), but the crux of the assignment and the distribution of points will not change. Any there any changes to the assignment, all changes will be document in the “Change History” section of this assignment,

1 Assignment Description

The goal of this assignment is for you to learn how to convert between numbering systems and how binary operations work. A secondary goal is for you to continue to develop your programming skills. This program should be developed in Java version 8, and your classes must reside in the package “cs250.hw1”.

Your program will be provided with 3 arguments at the command line. You are required to perform a set of operations over these arguments. These operations have been broken up into tasks: the points distribution for each task and the restrictions (and accompanying deductions) are specified in the grading section of this assignment.

We will compile and run your program using the commands below, replacing “<num1>”, “<num2>”, and “<num3>” with numbers as described in the tasks. See the “Example Outputs” section for more examples on how your program will be run. Your code must compile and run on the Linux machines in the CSB 120 computer lab, using Java version 8 (the default on these computers). It is highly recommended to develop your code using VS Code, either directly on the machines or through the Remote SSH extension in VS Code. Other IDE’s or text editors may not be supported by the instructors.

Command line compilation: javac cs250/hw1/Operations.java

Command line execution format: java cs250.hw1.Operations <num1> <num2> <num3>

Command line execution example: java cs250.hw1.Operations 0x123 0b111 17